moved all call-related definitions to primitives
