+++ primitives/header-chain/src/lib.rs
@@ -24,8 +24,8 @@ use crate::justification::{
 	GrandpaJustification, JustificationVerificationContext, JustificationVerificationError,
 };
 use bp_runtime::{
-	BasicOperatingMode, Chain, HashOf, HasherOf, HeaderOf, StorageProofError,
-	UnderlyingChainProvider, UnverifiedStorageProof, VerifiedStorageProof,
+	BasicOperatingMode, Chain, HashOf, HasherOf, StorageProofError, UnderlyingChainProvider,
+	UnverifiedStorageProof, VerifiedStorageProof,
 };
 use codec::{Codec, Decode, Encode, EncodeLike, MaxEncodedLen};
 use core::{clone::Clone, cmp::Eq, default::Default, fmt::Debug};
@@ -35,6 +35,10 @@ use serde::{Deserialize, Serialize};
 use sp_consensus_grandpa::{AuthorityList, ConsensusLog, SetId, GRANDPA_ENGINE_ID};
 use sp_runtime::{traits::Header as HeaderT, Digest, RuntimeDebug};
 use sp_std::{boxed::Box, vec::Vec};
+
+pub use call_info::{BridgeGrandpaCall, BridgeGrandpaCallOf, SubmitFinalityProofInfo};
+
+mod call_info;
 
 pub mod justification;
 pub mod storage_keys;
@@ -239,29 +243,6 @@ pub trait FindEquivocations<FinalityProof, FinalityVerificationContext, Equivoca
 		source_proofs: &[FinalityProof],
 	) -> Result<Vec<EquivocationProof>, Self::Error>;
 }
-
-/// A minimized version of `pallet-bridge-grandpa::Call` that can be used without a runtime.
-#[derive(Encode, Decode, Debug, PartialEq, Eq, Clone, TypeInfo)]
-#[allow(non_camel_case_types)]
-pub enum BridgeGrandpaCall<Header: HeaderT> {
-	/// `pallet-bridge-grandpa::Call::submit_finality_proof`
-	#[codec(index = 0)]
-	submit_finality_proof {
-		/// The header that we are going to finalize.
-		finality_target: Box<Header>,
-		/// Finality justification for the `finality_target`.
-		justification: justification::GrandpaJustification<Header>,
-	},
-	/// `pallet-bridge-grandpa::Call::initialize`
-	#[codec(index = 1)]
-	initialize {
-		/// All data, required to initialize the pallet.
-		init_data: InitializationData<Header>,
-	},
-}
-
-/// The `BridgeGrandpaCall` used by a chain.
-pub type BridgeGrandpaCallOf<C> = BridgeGrandpaCall<HeaderOf<C>>;
 
 /// Substrate-based chain that is using direct GRANDPA finality.
 ///
