+++ primitives/parachains/src/lib.rs
@@ -22,7 +22,7 @@
 pub use bp_header_chain::StoredHeaderData;
 
 use bp_polkadot_core::{
-	parachains::{ParaHash, ParaHead, ParaHeadsProof, ParaId},
+	parachains::{ParaHash, ParaHead, ParaId},
 	BlockNumber as RelayBlockNumber, Hash as RelayBlockHash,
 };
 use bp_runtime::{
@@ -35,6 +35,10 @@ use scale_info::TypeInfo;
 use sp_core::storage::StorageKey;
 use sp_runtime::{traits::Header as HeaderT, RuntimeDebug};
 use sp_std::{marker::PhantomData, prelude::*};
+
+pub use call_info::{BridgeParachainCall, SubmitParachainHeadsInfo};
+
+mod call_info;
 
 /// Best known parachain head hash.
 #[derive(Clone, Decode, Encode, MaxEncodedLen, PartialEq, RuntimeDebug, TypeInfo)]
@@ -165,20 +169,4 @@ impl ParaStoredHeaderDataBuilder for C {
 
 		None
 	}
-}
-
-/// A minimized version of `pallet-bridge-parachains::Call` that can be used without a runtime.
-#[derive(Encode, Decode, Debug, PartialEq, Eq, Clone, TypeInfo)]
-#[allow(non_camel_case_types)]
-pub enum BridgeParachainCall {
-	/// `pallet-bridge-parachains::Call::submit_parachain_heads`
-	#[codec(index = 0)]
-	submit_parachain_heads {
-		/// Relay chain block, for which we have submitted the `parachain_heads_proof`.
-		at_relay_block: (RelayBlockNumber, RelayBlockHash),
-		/// Parachain identifiers and their head hashes.
-		parachains: Vec<(ParaId, ParaHash)>,
-		/// Parachain heads proof.
-		parachain_heads_proof: ParaHeadsProof,
-	},
 }
